move AKS diagnostics commands to their own file
